requirements.txt
Django==3.1.1
djangorestframework==3.13.1
mysqlclient==2.0.1
django-mysql==3.8.1
django-model-utils==4.2.0